move entities and insights to recoil
