+++ web_frontend/src/recoil.ts
@@ -1,4 +1,5 @@
 import { atom } from "recoil";
+import { Entity, Insight } from "./types";
 
 export const isExplicitListeningState = atom<boolean>({
   key: "isExplicitListening",
@@ -15,4 +16,11 @@ export const isExplicitListeningState = atom<boolean>({
 export const isRecognizingState = atom<boolean>({
   key: "isRecognizing",
   default: true,
+});
+
+export const entitiesState = atom<Entity[]>({ key: "entities", default: [] });
+
+export const explicitInsightsState = atom<Insight[]>({
+  key: "explicitInsights",
+  default: [],
 });
